# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FanHook API",
    "version": "0.1.0",
    "description": "Webhook fanout and relay API. Route incoming webhooks from Stripe, GitHub, and more to multiple downstream destinations with automatic retries."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key (UUID)"
      }
    },
    "schemas": {
      "Sink": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "provider": {
            "type": "string",
            "enum": ["stripe", "github", "generic"],
            "default": "generic"
          },
          "api_key": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Route": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sink_id": { "type": "string", "format": "uuid" },
          "url": { "type": "string", "format": "uri" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sink_id": { "type": "string", "format": "uuid" },
          "provider": { "type": "string" },
          "payload": { "type": "string", "description": "Raw JSON payload string" },
          "received_at": { "type": "string", "format": "date-time" },
          "status": {
            "type": "string",
            "enum": ["pending", "delivered", "failed"]
          },
          "delivery_attempts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DeliveryAttempt" }
          }
        }
      },
      "DeliveryAttempt": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "event_id": { "type": "string", "format": "uuid" },
          "route_id": { "type": "string", "format": "uuid" },
          "attempt_number": { "type": "integer", "minimum": 1, "maximum": 3 },
          "status": {
            "type": "string",
            "enum": ["success", "failed"]
          },
          "http_status": { "type": "integer", "nullable": true },
          "attempted_at": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/api/sinks": {
      "post": {
        "summary": "Create a new sink",
        "operationId": "createSink",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "provider": {
                    "type": "string",
                    "enum": ["stripe", "github", "generic"],
                    "default": "generic"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sink created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sink_id": { "type": "string", "format": "uuid" },
                    "ingest_url": { "type": "string" },
                    "api_key": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List sinks for the authenticated API key",
        "operationId": "listSinks",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of sinks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Sink" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/sinks/{sinkId}/events": {
      "get": {
        "summary": "Get last 50 events for a sink",
        "operationId": "listEvents",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sinkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of events with nested delivery attempts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Event" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/sinks/{sinkId}/routes": {
      "post": {
        "summary": "Add a destination route to a sink",
        "operationId": "createRoute",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sinkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Route" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/sinks/{sinkId}/routes/{routeId}": {
      "delete": {
        "summary": "Delete a route from a sink",
        "operationId": "deleteRoute",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sinkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "routeId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Route deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Route not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/ingest/{sinkId}": {
      "post": {
        "summary": "Receive an incoming webhook for fanout",
        "operationId": "ingestWebhook",
        "security": [],
        "parameters": [
          {
            "name": "sinkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Any valid JSON webhook payload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook received and queued for fanout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": { "type": "boolean" },
                    "routed": { "type": "integer" },
                    "event_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Signature verification failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Sink not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  }
}